backend/requirements.txt
Flask
google-cloud-speech
google-cloud-translate
spotipy
openai
